\name{getVignette}
\alias{getVignette}
\title{Create a copy of the SGPreports Vignette Files in the Working Directory}
\usage{
getVignette())
}
\description{
Creates a directory called SGPreports_Vignette in the working directory and makes it the working directory.  The file folder contains all necessary .Rmd source files, JavaScript, CSS stylesheets, images, etc needed to produce exemplar documents with the main functions of the SGPreports package.
}
